data/requirements.txt
python-dotenv
pymongo
pinecone-client
google-generativeai
ijson
